build: change some classes
