+++ .github/ISSUE_TEMPLATE/question.md
@@ -18,7 +18,7 @@ If applicable, add screenshots to help explain your problem.
 
 **Desktop (please complete the following information):**
  - OS: [e.g. Ubuntu]
- - Version [e.g. 22]
+ - Version [e.g. 0.1.7]
 
 **Additional context**
 Add any other context about the problem here.
