+++ .github/pull_request_template.md
@@ -5,58 +5,21 @@
 
 Issue number : *add reference to issue number (if applicable)*
 
-If you filled out the issue number move on to the checklist sectionm if not fill in these deatils:
-
-**Description of functionality**
-*Add a bit of infomation about what you intend for this feature to do*
-
-**Current behaviour**
-A clear and concise description of what currently happens.
-
-**Desired behavior**
-A clear and concise description of what you want to happen.
-
-**Desktop (please complete the following information):**
- - OS: [e.g. iOS]
- - Version [e.g. 22]
+**Description**
+- What is the feature?
 
 ---
 
 # Bug Fix
-
-## Details
-
 *Fill out below fields*
 
 Issue number : *add reference to issue number (if applicable)*
 
-If you filled out the issue number move on to the checklist sectionm if not fill in these deatils:
-
-**Describe the bug**
-A clear and concise description of what the bug is.
-
-**To Reproduce**
-Steps to reproduce the behavior:
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
-
-**Current behaviour**
-A clear and concise description of what currently happens.
-
-**Expected behavior**
-A clear and concise description of what you expected to happen.
-
-**Desktop (please complete the following information):**
- - OS: [e.g. iOS]
- - Version [e.g. 22]
+**Description**
+- What is the bug?
 
 ---
 
 # Checklist
 
-- [ ] Commenting is present in all Classes, modules, and functions.
-- [ ] Functionality has been validated on Linux/Mac
-- [ ] Functionality has been validated on Windows
-- [ ] Validated this change does not break backwards compatibility without good reason
+- [ ] Passing tests
