+++ docs/changelog/index.md
@@ -1,1 +1,5 @@
 # CHANGELOG
+
+## v0.1.7
+- Add a new feature: `injector`.
+- Add more base classes.
