+++ docs/mkdocs.yml
@@ -128,6 +128,8 @@ nav:
       - API: user-guide/api.md
       - Components: user-guide/component.md
       - Core: user-guide/core.md
+      - Model: user-guide/model.md
+      - Dataset: user-guide/dataset.md
   - Benchmark:
       - benchmark/index.md
