+++ docs/user-guide/api.md
@@ -1,1 +1,3 @@
 # API Overview
+
+## opts
