+++ docs/user-guide/index.md
@@ -1,1 +1,5 @@
 # User Guide Overview
+
+## Architecture
+
+## Docs structure
