+++ fedhf/component/trainer/fedasync_trainer.py
@@ -13,8 +13,6 @@ import time
 from copy import deepcopy
 
 from tqdm import tqdm
-from fedhf.component.logger import Logger
-from fedhf.model import build_criterion, build_optimizer
 from .base_trainer import BaseTrainer
