+++ fedhf/core/__init__.py
@@ -20,6 +20,9 @@ __all__ += ["SimulatedSyncCoordinator", "SimulatedAsyncCoordinator"]
 
 __all__ += ["server_factory", "client_factory", "coordinator_factory"]
 
+__all__ += ["Injector"]
+
 from .server import build_server, SimulatedServer, server_factory
 from .client import build_client, SimulatedClient, client_factory
 from .coordinator import build_coordinator, SimulatedSyncCoordinator, SimulatedAsyncCoordinator, coordinator_factory
+from .injector import Injector
