+++ fedhf/core/client/base_client.py
@@ -8,10 +8,31 @@
 @License :   Apache License 2.0
 """
 
-
 from abc import ABC, abstractmethod
 
+from fedhf.api import Logger
+from fedhf.component import Evaluator, build_trainer
+from fedhf.model import build_criterion, build_model, build_optimizer
 
-class BaseClient(ABC):
+
+class AbsClient(ABC):
     def __init__(self) -> None:
         super().__init__()
+
+    @abstractmethod
+    def train(self):
+        raise NotImplementedError
+
+    def evalute(self):
+        raise NotImplementedError
+
+
+class BaseClient(AbsClient):
+    def __init__(self, args, client_id) -> None:
+        self.args = args
+        self.client_id = client_id
+
+        self.trainer = build_trainer(self.args.trainer)(self.args)
+        self.evaluator = Evaluator(args)
+
+        self.logger = Logger(args)
