+++ fedhf/core/client/simulated_client.py
@@ -11,7 +11,7 @@
 import re
 from torch.utils.data import DataLoader
 
-from fedhf.component import Evaluator, Logger, build_trainer
+from fedhf.component import Evaluator, build_trainer
 from fedhf.model import build_criterion, build_model, build_optimizer
 
 from .base_client import BaseClient
@@ -19,13 +19,7 @@ from .base_client import BaseClient
 
 class SimulatedClient(BaseClient):
     def __init__(self, args, client_id) -> None:
-        self.args = args
-        self.client_id = client_id
-
-        self.trainer = build_trainer(self.args.trainer)(self.args)
-        self.evaluator = Evaluator(args)
-
-        self.logger = Logger(args)
+        super(SimulatedClient, self).__init__(args, client_id)
 
     def train(self, data, model, device='cpu'):
         dataloader = DataLoader(data, batch_size=self.args.batch_size)
