+++ fedhf/core/coordinator/simulated_async_coordinator.py
@@ -15,7 +15,7 @@ import numpy as np
 
 from fedhf.core import build_server, build_client
 
-from fedhf.component import Logger, build_sampler
+from fedhf.component import build_sampler
 from fedhf.dataset import ClientDataset, build_dataset
 
 from .base_coordinator import SimulatedBaseCoordinator
